feat: 实现 Host 端单元测试框架及全部模块测试用例
- 新增 tests/ 目录，包含轻量级测试框架、mock 硬件层
- 实现 12 个模块的单元测试：heap4, task, scheduler, queue,
  semaphore, mutex, timer, eventgroup, mempool, mailbox, notify, watchdog
- 共 442 个测试用例，覆盖创建/销毁、参数校验、状态查询等场景
- 修复 mutex unlock 未清除 owner 的 bug
- 修复 heap4/mempool 在 64 位平台的指针截断问题
- 为 kernel 代码添加 TEST_HOST_BUILD 条件编译，屏蔽 ARM 汇编
- 主 Makefile 添加 Windows 兼容性支持

+++ kernel/heap4.c
@@ -10,6 +10,7 @@
 #include "heap4.h"
 #include "os_config.h"
 #include <string.h>
+#include <stdint.h>
 
 /* ========== Internal Data ========== */
 
@@ -125,7 +126,7 @@ void os_heap_init(void)
     uint8_t *aligned_heap;
 
     /* Align the heap start address */
-    aligned_heap = (uint8_t*)ALIGN_UP((uint32_t)uc_heap);
+    aligned_heap = (uint8_t*)ALIGN_UP((uintptr_t)uc_heap);
 
     /* Initialize the first block spanning the entire heap */
     first_block = (os_block_header_t*)aligned_heap;
